Refactor derived markdown rendering modules
